Fixes for existing commands, better help system (#124)
* Remove some broken pic links

* Fix ClearCommand

* Fix FuzzySearch trailing no result output

* Remove Lua command

* Fix MAL related code

Allow selfhosters to access MAL by adding proper credential options.
Check MAL credentials on startup.
Due to MAL API being wonky, reduce http timeouts and run the requests async for a cleaner user experience.
Introduce a global executor service for small async tasks.

* Fix #8 say command not displaying newlines

* Add Discord terminology compliant aliases to server- and userinfo comms

* Fix skip in range command on several levels

* Extend HelpCommand for better user experience

some minor fixes happened too, like showing 5 tracks for selection as originally intended

* Introduce testing with JDA access and add test for command help function

Basically, this commit paves the way for writing live tests for the bot.

* Fix a rare™ NullPointerException

* Fix #112 CatgirlCommand by bypassing CloudFlare js challenge

* High quality for development distribution

Motivated by the changes to scope/distribution handling. Using the development distribution potential selfhosters can host a bot that run both main and music commands, and play high quality music.

* Add CommandsCommand, redo MusicHelpCommand, categorize all commands

* Fix some newly added commands to comply with current standards

* Fix #113 Pat, Roll and Facedesk commands

* Add imgur album support for RandomImageCommand

* Rehost RemoteFileCommand links on imgur, remove broken/lost ones

* Rehost images on imgur

* Fix tests; Add maven integration testing plugin

* Hide live testing for now

* Restrict i18n help strings to music, moderation and utility commands

And move the usage and formatting clutter to the code

* Set imgur refresh to a production value

* Move some clutter from the language strings to the code

* Update lang and TestCommand help

* Update lang

+++ FredBoat/src/main/java/fredboat/audio/MusicPersistenceHandler.java
@@ -146,13 +146,15 @@ public class MusicPersistenceHandler {
         if(Config.CONFIG.getDistribution() == DistributionEnum.MUSIC) {
             log.warn("Music persistence loading is currently disabled!");
 
-            for (File f : dir.listFiles()) {
-                boolean deleted = f.delete();
-                log.info(deleted ? "Deleted persistence file: " + f : "Failed to delete persistence file: " + f);
+            File[] files = dir.listFiles();
+            if (files != null) {
+                for (File f : files) {
+                    boolean deleted = f.delete();
+                    log.info(deleted ? "Deleted persistence file: " + f : "Failed to delete persistence file: " + f);
+                }
+
+                dir.delete();
             }
-
-            dir.delete();
-
             return;
         }
